Improvements to automate tasks using pyrogram
When a user has automated reading of chats/channels/groups content or has the message handler active, the links to the channels usually also appear in the following format: "https://t.me/telegram". Right now, we get a 400 error if we want to authenticate using the "https://t.me/telegram" link.

With this improvement in line 70, the user can use either just the public channel nickname or the full link. It makes it easier for the user to use any type of link and he will have no problem accessing it.

With this change, apart from improving the usability of the library, some issues such as "https://github.com/pyrogram/pyrogram/issues/1077" from August 2022 are also fixed.